Refactor search filters to better allow switching between D1 and D2
